feat : add test

+++ src/main/java/com/app/Notes.java
@@ -3,13 +3,13 @@ package com.app;
 import java.time.Instant;
 
 public class Notes {
-    private int valeur;
+    private double valeur;
     private Examens examen;
     private Instant dateChangement;
     private String description;
     private Etudiants etudiant;
 
-    public Notes(int valeur, Examens examen, String description, Instant dateChangement, Etudiants etudiant) {
+    public Notes(double valeur, Examens examen, String description, Instant dateChangement, Etudiants etudiant) {
         this.valeur = valeur;
         this.examen = examen;
         this.description = description;
@@ -17,7 +17,7 @@ public class Notes {
         this.etudiant = etudiant;
     }
 
-    public int getValeur() {
+    public double getValeur() {
         return valeur;
     }
 
@@ -28,4 +28,5 @@ public class Notes {
     public Examens getExamen() {
         return examen;
     }
+
 }
